[refactor] Used a native input elements for the chat message composition box and search box (30m)
I'm unsure why the enter event wasn't firing when in Members, but ran across a couple-years-old blog post wherein Robert Jackson (a mainstay of the Ember community and part of the core team) mentions that going forward, the Input/Textarea helpers should only be used if there's no backing class for a component; if there's a backing class, you can and should go ahead and use the native elements. So that's what we'll do, since it seems to work. 😅